requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
boltons==25.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
dataclasses-json==0.6.7
frozenlist==1.5.0
fs-pochta-api==1.2.1
gigachat==0.1.39.post1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
isodate==0.7.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.47
langchain-gigachat==0.3.8
langchain-text-splitters==0.3.7
langgraph==0.3.18
langgraph-checkpoint==2.0.21
langgraph-prebuilt==0.1.4
langgraph-sdk==0.1.58
langsmith==0.3.18
lxml==5.3.1
marshmallow==3.26.1
msgpack==1.1.0
multidict==6.2.0
mypy-extensions==1.0.0
numpy==2.2.4
orjson==3.10.15
packaging==24.2
platformdirs==4.3.7
propcache==0.3.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dotenv==1.0.1
python-telegram-bot==22.0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.39
suds==1.2.0
tenacity==9.0.0
tiktoken==0.9.0
types-requests==2.32.0.20250306
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
zeep==4.3.1
zstandard==0.23.0